* Prototype error

+++ src/pl-funcs.h
@@ -46,7 +46,7 @@ COMMON(void) 		cleanupMemAlloc(void);
 COMMON(void) 		mergeAllocPool(AllocPool to, AllocPool from);
 COMMON(word) 		globalFunctor(functor_t def);
 COMMON(word) 		globalString(long len, const char *s);
-COMMON(word) 		globalWString(long len, const pl_wchar_t *s);
+COMMON(word) 		globalWString(unsigned len, const pl_wchar_t *s);
 COMMON(char *)		getCharsString__LD(word w, unsigned *len ARG_LD);
 COMMON(pl_wchar_t *)	getCharsWString__LD(word w, unsigned *len ARG_LD);
 COMMON(Word) 		newTerm(void);
